ui_slider1

+++ ESP_Hover_Browser_GY_521/ESP_Hover_Browser_GY_521.ino
@@ -522,7 +522,7 @@ void handleSlider1(int value)
   DEBUG_SERIAL.println(value);
 #endif
 
-  ui_slider1 = (float)value;
+  ui_slider1 = value;
 
   updateMotors();
 }
